add error delete

+++ src/routes/deletePokemon.js
@@ -3,6 +3,11 @@ const { Pokemon } = require('../db/sequelize')
 module.exports = (app) => {
   app.delete('/api/pokemons/:id', (req, res) => {
     Pokemon.findByPk(req.params.id).then(pokemon => {
+      if(pokemon === null){
+        const message = 'Le pokémon demandé n\'existe pas. Réessayez avce un autre identifiant.'
+        return res.status(404).json({message})
+      }
+
       const pokemonDeleted = pokemon;
       Pokemon.destroy({
         where: { id: pokemon.id }
@@ -11,6 +16,10 @@ module.exports = (app) => {
         const message = `Le pokémon avec l'identifiant n°${pokemonDeleted.id} a bien été supprimé.`
         res.json({message, data: pokemonDeleted })
       })
+      .catch(error => {
+        const message = 'Le pokemon n\'a pas pu être supprimé. Réessayez dans quelques instants.';
+        res.status(500).json({ message, data: error})
+        })
     })
   })
 }
